+++ server_messenger/Accounts/src/UserAccount.cpp
@@ -269,7 +269,7 @@ void UserAccount::create_solo_chat(error_code& ec)
     size_t length = socket_->read_some(asio::buffer(buf), ec);
     if (checkError(ec)) return;
 
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
         if (!isdigit(buf[i])) {
             acDEBUG_LOG("DEBUG_Temp_account", "not a number entered");
 
@@ -291,7 +291,7 @@ void UserAccount::add_solo_chat(error_code& ec)
     size_t length = socket_->read_some(asio::buffer(buf), ec);
     if (checkError(ec)) return;
 
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
         if (!isdigit(buf[i])) {
             acDEBUG_LOG("DEBUG_Temp_account", "not a number entered");
 
@@ -326,7 +326,7 @@ void UserAccount::create_group_chat(error_code& ec)
     size_t length = socket_->read_some(asio::buffer(buf), ec);
     if (checkError(ec)) return;
 
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
         if (!isdigit(buf[i])) {
             acDEBUG_LOG("DEBUG_Temp_account", "not a number entered");
 
@@ -387,7 +387,7 @@ void UserAccount::invite_user(error_code& ec)
     size_t length = socket_->read_some(asio::buffer(buf), ec);
     if (checkError(ec)) return;
 
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
         if (!isdigit(buf[i])) {
             acDEBUG_LOG("DEBUG_Temp_account", "not a number entered");
 
@@ -421,7 +421,7 @@ void UserAccount::add_user_group_chat(error_code& ec)
     size_t length = socket_->read_some(asio::buffer(buf), ec);
     if (checkError(ec)) return;
 
-    for (int i = 0; i < length; i++) {
+    for (size_t i = 0; i < length; i++) {
         if (!isdigit(buf[i])) {
             acDEBUG_LOG("DEBUG_Temp_account", "not a number entered");
